move lib/ui into vscode/webviews, remove lib/ui
The `@sourcegraph/cody-ui` package (`lib/ui`) abstraction has not worked. We may need to add something like it back, but for now, let's remove it so we can reduce the complexity of the code.
