--- lib/ui/.gitignore
@@ -1,3 +1,0 @@
-node_modules/
-out/
-dist/
