--- lib/ui/README.md
@@ -1,3 +1,0 @@
-# Cody UI shared library
-
-The `@sourcegraph/cody-ui` package contains UI code that is shared among Cody clients.
