--- lib/ui/src/globals.d.ts
@@ -1,4 +1,0 @@
-declare module '*.module.css' {
-    const classes: { readonly [key: string]: string }
-    export default classes
-}
