--- lib/ui/src/index.css
@@ -1,1 +1,0 @@
-@import url(./utils/highlight.css);
