--- lib/ui/tsconfig.json
@@ -1,17 +1,0 @@
-{
-  "extends": "../../tsconfig.json",
-  "compilerOptions": {
-    "module": "ESNext",
-    "rootDir": "src",
-    "outDir": "dist",
-    "jsx": "react-jsx",
-    "lib": ["ESNext", "DOM", "DOM.Iterable"],
-  },
-  "include": ["src"],
-  "exclude": ["dist"],
-  "references": [
-    {
-      "path": "../shared",
-    },
-  ],
-}
