+++ vscode/tsconfig.json
@@ -5,7 +5,7 @@
     "rootDir": ".",
     "outDir": "dist/tsc",
     "jsx": "react-jsx",
-    "lib": ["ESNext", "DOM"]
+    "lib": ["ESNext", "DOM", "DOM.Iterable"]
   },
   "include": [
     "src",
@@ -29,8 +29,5 @@
     {
       "path": "../lib/shared"
     },
-    {
-      "path": "../lib/ui"
-    }
   ]
 }
