+++ vscode/webviews/App.tsx
@@ -12,8 +12,8 @@ import {
     type ModelProvider,
     type TranscriptJSON,
 } from '@sourcegraph/cody-shared'
-import type { UserAccountInfo } from '@sourcegraph/cody-ui/src/Chat'
-import { EnhancedContextEnabled } from '@sourcegraph/cody-ui/src/chat/components/EnhancedContext'
+import type { UserAccountInfo } from './LibChat'
+import { EnhancedContextEnabled } from './chat/components/EnhancedContext'
 
 import type { AuthMethod, AuthStatus, LocalEnv } from '../src/chat/protocol'
