+++ vscode/webviews/chat/BlinkingCursor.module.css
