+++ vscode/webviews/chat/BlinkingCursor.tsx
