+++ vscode/webviews/chat/ErrorItem.tsx
@@ -2,7 +2,7 @@ import React, { useCallback } from 'react'
 
 import { type ChatError, RateLimitError } from '@sourcegraph/cody-shared'
 
-import type { ApiPostMessage, ChatButtonProps, UserAccountInfo } from '../Chat'
+import type { ApiPostMessage, ChatButtonProps, UserAccountInfo } from '../LibChat'
 
 import styles from './ErrorItem.module.css'
