+++ vscode/webviews/chat/Transcript.story.tsx
