+++ vscode/webviews/chat/TranscriptItem.module.css
