+++ vscode/webviews/chat/actions/TranscriptAction.tsx
