+++ vscode/webviews/chat/components/EnhancedContext.tsx
