+++ vscode/webviews/icons/CodyLogo.tsx
