+++ vscode/webviews/icons/LLMProviderIcons.tsx
